UOELSA-1102: Korjaa tietotyyppi
- Muuta erikoistumiseen vaadittava tuntimäärä kokonaisluvusta liukuluvuksi vastaamaan opintooppaalle määritettyä tyyppiä

+++ src/main/kotlin/fi/elsapalvelu/elsa/domain/Teoriakoulutus.kt
@@ -37,7 +37,7 @@ data class Teoriakoulutus(
 
     @get: Min(value = 0)
     @Column(name = "erikoistumiseen_hyvaksyttava_tuntimaara")
-    var erikoistumiseenHyvaksyttavaTuntimaara: Int? = null,
+    var erikoistumiseenHyvaksyttavaTuntimaara: Double? = null,
 
     @OneToMany(
         mappedBy = "teoriakoulutus",
